use uint64 for TotalAssets, TotalAssetParms, TotalAppParams, TotalAppLocalStates (#3511)

+++ ledger/accountdb.go
@@ -1092,10 +1092,10 @@ type baseAccountData struct {
 	TotalAppSchemaNumUint      uint64            `codec:"f"`
 	TotalAppSchemaNumByteSlice uint64            `codec:"g"`
 	TotalExtraAppPages         uint32            `codec:"h"`
-	TotalAssetParams           uint32            `codec:"i"`
-	TotalAssets                uint32            `codec:"j"`
-	TotalAppParams             uint32            `codec:"k"`
-	TotalAppLocalStates        uint32            `codec:"l"`
+	TotalAssetParams           uint64            `codec:"i"`
+	TotalAssets                uint64            `codec:"j"`
+	TotalAppParams             uint64            `codec:"k"`
+	TotalAppLocalStates        uint64            `codec:"l"`
 
 	baseOnlineAccountData
 
@@ -1165,10 +1165,10 @@ func (ba *baseAccountData) SetAccountData(ad *basics.AccountData) {
 	ba.TotalAppSchemaNumUint = ad.TotalAppSchema.NumUint
 	ba.TotalAppSchemaNumByteSlice = ad.TotalAppSchema.NumByteSlice
 	ba.TotalExtraAppPages = ad.TotalExtraAppPages
-	ba.TotalAssetParams = uint32(len(ad.AssetParams))
-	ba.TotalAssets = uint32(len(ad.Assets))
-	ba.TotalAppParams = uint32(len(ad.AppParams))
-	ba.TotalAppLocalStates = uint32(len(ad.AppLocalStates))
+	ba.TotalAssetParams = uint64(len(ad.AssetParams))
+	ba.TotalAssets = uint64(len(ad.Assets))
+	ba.TotalAppParams = uint64(len(ad.AppParams))
+	ba.TotalAppLocalStates = uint64(len(ad.AppLocalStates))
 }
 
 func (ba *baseAccountData) GetLedgerCoreAccountData() ledgercore.AccountData {
@@ -2731,7 +2731,7 @@ func (iterator *encodedAccountsBatchIter) Next(ctx context.Context, tx *sql.Tx, 
 		return nil
 	}
 
-	var totalAppParams, totalAppLocalStates, totalAssetParams, totalAssets uint32
+	var totalAppParams, totalAppLocalStates, totalAssetParams, totalAssets uint64
 	// emptyCount := 0
 	resCb := func(addr basics.Address, cidx basics.CreatableIndex, ctype basics.CreatableType, resData *resourcesData, encodedResourceData []byte) error {
 		emptyBaseAcct := baseAcct.TotalAppParams == 0 && baseAcct.TotalAppLocalStates == 0 && baseAcct.TotalAssetParams == 0 && baseAcct.TotalAssets == 0
@@ -3007,7 +3007,7 @@ func LoadAllFullAccounts(
 		return nil
 	}
 
-	var totalAppParams, totalAppLocalStates, totalAssetParams, totalAssets uint32
+	var totalAppParams, totalAppLocalStates, totalAssetParams, totalAssets uint64
 	resCb := func(addr basics.Address, cidx basics.CreatableIndex, ctype basics.CreatableType, resData *resourcesData, encodedResourceData []byte) error {
 		if resData != nil {
 			// resData can be nil if a base account does not have resources
